+++ WenigerTorben-Bot/Audio/Queueing/AudioRequest.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
 using Discord;
-using WenigerTorbenBot.Audio.AudioSource;
+using WenigerTorbenBot.Audio.Source;
 
 namespace WenigerTorbenBot.Audio.Queueing
 {
