+++ WenigerTorben-Bot/Modules/Text/AudioModule.cs
@@ -10,9 +10,9 @@ using Discord;
 using Discord.Commands;
 using Discord.WebSocket;
 using Serilog;
-using WenigerTorbenBot.Audio.AudioSource;
 using WenigerTorbenBot.Audio.Queueing;
 using WenigerTorbenBot.Audio.Session;
+using WenigerTorbenBot.Audio.Source;
 using WenigerTorbenBot.Metadata;
 using WenigerTorbenBot.Services.Audio;
 using WenigerTorbenBot.Services.FFmpeg;
